requirements.txt
Django==1.10
mysqlclient
django-dotenv
xlrd
djangorestframework
